latest/requirements.txt
google-generativeai>=0.3.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
xgboost>=1.7.0
arch>=5.3.0
scipy>=1.10.0
